requirements.txt
sentencepiece==0.1.94
pytesseract
opencv-python
pillow
argostranslate